Route sine & cosine from IQ-module to ASG module for external phase control
For some applications, it might be helpful to control the phase of the AWG signal from the IQ module. While the IQ module allows generating a sine-wave which can be routed to the DACs, maybe we want to generate a different signal (e.g. square-wave) on the DACs, with a phase controlled by the IQ-module. This change allows for that.