Disable PID module and remove related imports due to resource constraints on FPGA. Started removing lockbox functionalities too
The FPGA resources are very close to being completely used. We have to remove some FPGA functionalities to make room for new features. As such, the PID module is temporarily disabled. We might however need it again in the future, e.g. for locking to the NV-resonances. Then we will have to remove other functionalities instead / or make the resource usage more efficient.

Further I have started removing the lockbox functionalities. I would like to go from the pyrpl code to a project more focussed on NV-ODMR functionalities. In the far future, I want to have a project completely focussed only on NV-ODMR things.