Add frequency control to PWM module and expand experimental control
For some experiments we want to not only change the pulse-width-modulation duty-cycle, but also use different PWM frequencies. (We have an experiment in that we communicate some values to a different device, using changing PWM frequencies).
Currently, the PWM module is not connected to the signal bus. This necessitates setting the PWM frequencies via the AMS module and routing them to the PWM module.
This is quite unelegant and will hopefully be changed in the future.