Add 3-Component FM Sine Generator (Fgen3) module and associated GUI
Created a new hardware module that can generate a 3-tone/3-frequency signal on the two output channels. The individual frequency components can be frequency modulated. The phase relation between the two channels can be controlled, to allow for single-sideband-mixing/IQ-mixing.
For the FM, the sine signal from the IQ0 module is used as a reference/as the modulating signal. This requires routing the signal from the IQ0 module via the DSP module, via the top module to the new fgen3 module.
It may be more elegant to include the fgen3 module in the DSP module, maybe will be done in the future (most likely not time for that lol).